requirements.txt
pytorch==0.3.1
numpy
scikit-learn
torchwordemb
tqdm
matplotlib
torchvision
Pillow
torchwordemb

